It was all a typo... Life is not Daijoubu

+++ core/src/me/rainma22/gdx/GameObject2D.java
@@ -4,25 +4,25 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Rectangle;
 
 public class GameObject2D {
-	int x,y,z;
+	float x,y,z;
 	TextureRegion texture;
-	int width,height;
+	float width,height;
 	int xmax,ymax;
 	int xcounter,ycounter;
 	private Rectangle rect;
 	int fps;
 	float scale;
 	float rotation;
-	public GameObject2D(int x,int y,TextureRegion t,int width,int height,int fps,float scale) {
+	public GameObject2D(float x,float y,TextureRegion t,float width,float height,int fps,float scale) {
 		this.x=x;this.y=y;texture=t; this.width=width;this.height=height;
-		xmax=t.getRegionWidth()/width;ymax=t.getRegionHeight()/height;
+		xmax=(int)(t.getRegionWidth()/width);ymax=(int)(t.getRegionHeight()/height);
 		xcounter=0;ycounter=0;rect=new Rectangle(x, y, width*scale, height*scale);
 		this.fps=fps;
 		this.scale=scale;
 	}
-	public GameObject2D(int x,int y,TextureRegion t,int width,int height,int fps,float scale,int beginCounter) {
+	public GameObject2D(float x,float y,TextureRegion t,float width,float height,int fps,float scale,int beginCounter) {
 		this.x=x;this.y=y;texture=t; this.width=width;this.height=height;
-		xmax=t.getRegionWidth()/width;ymax=t.getRegionHeight()/height;
+		xmax=(int)(t.getRegionWidth()/width);ymax=(int)(t.getRegionHeight()/height);
 		xcounter=0;ycounter=0;rect=new Rectangle(x, y, width*scale, height*scale);
 		this.fps=fps;
 		this.scale=scale;
@@ -40,7 +40,7 @@ public class GameObject2D {
 		this.rotation=rotation;
 	}
 	public TextureRegion getNextTexture(long i) {
-		TextureRegion tr=new TextureRegion(texture, (width*xcounter), (height*ycounter), width, height);
+		TextureRegion tr=new TextureRegion(texture, (int)(width*xcounter), (int)(height*ycounter), (int)width, (int)height);
 		if((double)i %(60d/fps)==0)
 			xcounter++;
 		if (xcounter>=xmax) {
@@ -51,7 +51,7 @@ public class GameObject2D {
 		return(tr);
 	}
 	public TextureRegion getPreviousTexture(long i) {
-		TextureRegion tr=new TextureRegion(texture, (width*xcounter), (height*ycounter), width, height);
+		TextureRegion tr=new TextureRegion(texture, (int)(width*xcounter), (int)(height*ycounter), (int)width, (int)height);
 		if((double)i %(60d/fps)==0)
 			xcounter--;
 		if (xcounter<=0) {
@@ -62,13 +62,13 @@ public class GameObject2D {
 		}
 		return(tr);
 	}
-	public int getX() {
+	public float getX() {
 		return this.x;
 	}
-	public int getY() {
+	public float getY() {
 		return this.y;
 	}
-	public int getZ() {
+	public float getZ() {
 		return this.z;
 	}
 	public float getScale(){
